requirements.txt
fastapi
uvicorn
torch
transformers
faiss-cpu
numpy
pillow
pandas
tqdm
